feat: move pages within src
