+++ src/pages/api/revalidate.ts
@@ -1,5 +1,5 @@
 import { NextApiRequest, NextApiResponse } from 'next';
-import { staticPaths } from '../../src/staticPaths';
+import { staticPaths } from '../../staticPaths';
 
 const handler = async (req: NextApiRequest, res: NextApiResponse) => {
   if (req.method !== 'POST') {
